Fix link failure with unresolved references in --debug build.
Root problem is recursive dependency between libdbug and libmysys. These
libraries are not meant to be built as .so.

Fix by making them into libtool convenience libraries instead. This avoids
introducing new .so that cause problems, and still allows to build a working
libmysqld.so.

Note that with this we no longer install libdbug, libmystrings, libmysys,
or libvio. This should not be a problem, as there should be no separate
programs using these libraries.